proposal_engine/requirements.txt
flask>=3.0
python-pptx>=1.0
Pillow>=10.0
